Revert "[fix] max_pool1d: composite compliance (#70900)"
This reverts commit b7222e15b6a457099b74420e29b3a39a3e8b5f1a.

We are conservatively reverting this because it broke a test in functorch.
The original PR added a `_max_pool1d_cpu` operator. I'm not sure if it
is actually safe to revert this due to the addition of the new operator
(someone may have serialized it between now and then) but because it has
only been two weeks this should be fine.

Test Plan:
- wait for tests

ghstack-source-id: 40a34bb586987566bf7495cd956a528b9ef2ecd4
Pull Request resolved: https://github.com/pytorch/pytorch/pull/71992