Add hook for functorch to error out with unoverridable autograd operations
I went through the manual_cpp_binding operations in
native_functions.yaml looking for important things that people use that
don't go through the dispatcher and came up with this.

There's currently no mechanism for functorch (or Tensor subclasses)
to change the behavior of tensor.requires_grad_() and
tensor.retains_grad() because these don't go through the dispatcher at
all.

This PR adds a hook for functorch to be able to throw an error on these.
In the future they should probably be overridable with torch_dispatch
(or at least configurable!).

ghstack-source-id: 088b8346eeacaa947791a023d06361b5becb7c08
Pull Request resolved: https://github.com/pytorch/pytorch/pull/72176