[Quant][core][devs] Changed empty_quantized call for quantized tensor to resize_output
Summary: This PR is part of a series of PRs addressing https://github.com/pytorch/pytorch/issues/54150,
related to using dispatcher for calls to quantized backends as opposed to if/else conditionals. This particular PR removes the call to empty_quantized for quantized tensors and substitutes
it for resize_output, which works for quantized tensors, based on current understanding.

ghstack-source-id: 5c435027aab1bd634cb7bec06e411781f7575714
Pull Request resolved: https://github.com/pytorch/pytorch/pull/72524