.github: Switch binary builds to use linux.2xlarge
Switches binary builds to use linux.2xlarge instead of linux.4xlarge.
The prior reason to using linux.4xlarge was because we were originally
hitting OOM issues. We were hitting those issues because we weren't
originally running binary_populate_env.sh but now that we are we can
safely return these to the more common machine type we use.

Signed-off-by: Eli Uriegas <eliuriegasfb.com>

ghstack-source-id: 3b7a6484a41a9ec00cb826772636e888dbb36def
Pull Request resolved: https://github.com/pytorch/pytorch/pull/73260