[PyTorch] [Model Tracer] Protect writes to set/map for capturing used operators, dtypes, custom classes, and build features
Pull Request resolved: https://github.com/pytorch/pytorch/pull/73327

The lack of a mutex to protect 4 static variables used by model tracer leads to a problem when multi-threaded/parallel operators are invoked when running a model. These cause the static variables to be read from and written to concurrently, corrupting their internal state.

Since tracing isn't a performance critical code path, it's okay to use a mutext to protect all writes and reads to these shared variables.

Verified that the problem to model tracer goes away after this change.

Differential Revision: [D34433840](https://our.internmc.facebook.com/intern/diff/D34433840/)

**NOTE FOR REVIEWERS**: This PR has internal Facebook specific changes or comments, please review them on [Phabricator](https://our.internmc.facebook.com/intern/diff/D34433840/)!
ghstack-source-id: 12229ce67855339892c90c2a9a76a4b9dfab0179